+++ randovania/game_description/pickup/standard_pickup.py
@@ -33,7 +33,7 @@ class StandardPickupDefinition(JsonDataclass):
     hide_from_gui: bool = dataclasses.field(default=False, metadata=EXCLUDE_DEFAULT)
     must_be_starting: bool = dataclasses.field(default=False, metadata=EXCLUDE_DEFAULT)
     original_location: PickupIndex | None = dataclasses.field(default=None, metadata=EXCLUDE_DEFAULT)
-    probability_offset: int = dataclasses.field(default=0, metadata=EXCLUDE_DEFAULT)
+    probability_offset: float = dataclasses.field(default=0.0, metadata=EXCLUDE_DEFAULT)
     probability_multiplier: float = dataclasses.field(default=1.0, metadata=EXCLUDE_DEFAULT)
     description: str | None = dataclasses.field(default=None, metadata=EXCLUDE_DEFAULT)
     extra: frozendict = dataclasses.field(default_factory=frozendict, metadata=EXCLUDE_DEFAULT)
