+++ randovania/games/cave_story/generator/hint_distributor.py
@@ -6,14 +6,15 @@ from randovania.game_description.db.node_identifier import NodeIdentifier
 from randovania.game_description.db.pickup_node import PickupNode
 from randovania.game_description.hint import HintItemPrecision, HintLocationPrecision, PrecisionPair
 from randovania.games.cave_story.layout.cs_configuration import CSConfiguration, CSObjective
-from randovania.generator.hint_distributor import HintDistributor, HintTargetPrecision, PreFillParams
+from randovania.generator.hint_distributor import HintDistributor, HintTargetPrecision
 
 if TYPE_CHECKING:
     from random import Random
 
     from randovania.game_description.game_patches import GamePatches
+    from randovania.generator.filler.filler_configuration import PlayerPool
     from randovania.generator.filler.player_state import PlayerState
-    from randovania.generator.filler.runner import PlayerPool
+    from randovania.generator.pre_fill_params import PreFillParams
 
 USE_GUARANTEED_HINTS = False
