+++ randovania/games/dread/generator/hint_distributor.py
@@ -5,14 +5,15 @@ from typing import TYPE_CHECKING
 from randovania.game_description.db.node_identifier import NodeIdentifier
 from randovania.game_description.hint import Hint, HintItemPrecision, HintLocationPrecision, HintType, PrecisionPair
 from randovania.games.dread.layout.dread_configuration import DreadConfiguration
-from randovania.generator.hint_distributor import HintDistributor, PreFillParams
+from randovania.generator.hint_distributor import HintDistributor
 
 if TYPE_CHECKING:
     from random import Random
 
     from randovania.game_description.game_patches import GamePatches
+    from randovania.generator.filler.filler_configuration import PlayerPool
     from randovania.generator.filler.player_state import PlayerState
-    from randovania.generator.filler.runner import PlayerPool
+    from randovania.generator.pre_fill_params import PreFillParams
 
 
 class DreadHintDistributor(HintDistributor):
