+++ randovania/games/prime2/generator/hint_distributor.py
@@ -13,7 +13,7 @@ from randovania.game_description.hint import (
 )
 from randovania.game_description.resources.pickup_index import PickupIndex
 from randovania.games.prime2.layout.echoes_configuration import EchoesConfiguration
-from randovania.generator.hint_distributor import HintDistributor, HintTargetPrecision, PreFillParams
+from randovania.generator.hint_distributor import HintDistributor, HintTargetPrecision
 from randovania.lib import enum_lib
 
 if TYPE_CHECKING:
@@ -21,8 +21,9 @@ if TYPE_CHECKING:
 
     from randovania.game_description.db.node_identifier import NodeIdentifier
     from randovania.game_description.game_patches import GamePatches
+    from randovania.generator.filler.filler_configuration import PlayerPool
     from randovania.generator.filler.player_state import PlayerState
-    from randovania.generator.filler.runner import PlayerPool
+    from randovania.generator.pre_fill_params import PreFillParams
 
 
 class EchoesHintDistributor(HintDistributor):
