+++ test/generator/filler/test_retcon.py
@@ -9,7 +9,7 @@ def test_select_weighted_action_weighted():
     rng = Random(1500)
     a = object()
     b = object()
-    actions = {a: 1, b: 2}
+    actions = {a: 1.0, b: 2.0}
 
     # Run
     r = retcon.select_weighted_action(rng, actions)
